api/endpoints/requirements.txt
fastapi
pydantic
langchain_openai
langchain_chroma
langchain_community
langchain_core
langchain
langchainhub
langsmith
langchain_huggingface
langchain-google-genai
pypdf
PyPDF2
python-docx
uvicorn
motor
python-dotenv
firebase-admin
boto3
chromadb
unstructured
python-multipart
aiofiles
psycopg2
debugpy